requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file


annotated-types==0.7.0
audioread==3.0.1
backports.functools-lru-cache==2.0.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
decorator==5.1.1
emoji==2.12.1
# en-core-web-lg==3.7.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
filelock==3.15.4
flask==3.0.3
fsspec==2024.6.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.23.5
hyperframe==5.2.0
idna==2.10
importlib-metadata==8.0.0
indic-transliteration==2.3.60
intel-openmp==2021.4.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language-data==1.2.0
lazy-loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
openai-whisper==20231117
packaging==24.1
pandas==2.2.2
pip==20.2.3
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
pydub==0.25.1
pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
roman==4.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
setuptools==49.2.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.0
tbb==2021.13.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
typer==0.12.3
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
